move telemetry to parent
