+++ cmd/applink-rp/main.go
@@ -19,9 +19,9 @@ import (
 	"github.com/project-radius/radius/pkg/linkrp/backend"
 	"github.com/project-radius/radius/pkg/linkrp/frontend"
 	"github.com/project-radius/radius/pkg/logging"
-	metricsservice "github.com/project-radius/radius/pkg/telemetry/metrics/service"
-	metricshostoptions "github.com/project-radius/radius/pkg/telemetry/metrics/service/hostoptions"
-	"github.com/project-radius/radius/pkg/telemetry/trace"
+	metricsservice "github.com/project-radius/radius/pkg/metrics/service"
+	metricshostoptions "github.com/project-radius/radius/pkg/metrics/service/hostoptions"
+	"github.com/project-radius/radius/pkg/trace"
 	"github.com/project-radius/radius/pkg/ucp/data"
 	"github.com/project-radius/radius/pkg/ucp/dataprovider"
 	"github.com/project-radius/radius/pkg/ucp/hosting"
