+++ cmd/rad/cmd/root.go
@@ -52,7 +52,7 @@ import (
 	"github.com/project-radius/radius/pkg/cli/output"
 	"github.com/project-radius/radius/pkg/cli/prompt"
 	"github.com/project-radius/radius/pkg/cli/setup"
-	"github.com/project-radius/radius/pkg/telemetry/trace"
+	"github.com/project-radius/radius/pkg/trace"
 
 	"github.com/spf13/cobra"
 	"github.com/spf13/viper"
