+++ cmd/ucpd/cmd/root.go
@@ -13,7 +13,7 @@ import (
 	"syscall"
 
 	"github.com/go-logr/logr"
-	"github.com/project-radius/radius/pkg/telemetry/trace"
+	"github.com/project-radius/radius/pkg/trace"
 	"github.com/project-radius/radius/pkg/ucp/server"
 	"github.com/project-radius/radius/pkg/ucp/ucplog"
 	"github.com/spf13/cobra"
