+++ pkg/metrics/service/hostoptions/hostoptions.go
@@ -7,7 +7,7 @@ package hostoptions
 
 import (
 	"github.com/project-radius/radius/pkg/armrpc/hostoptions"
-	"github.com/project-radius/radius/pkg/telemetry/metrics/provider"
+	"github.com/project-radius/radius/pkg/metrics/provider"
 )
 
 // HostOptions defines all of the settings that our metric's execution environment provides.
