+++ pkg/trace/options.go
