+++ pkg/ucp/server/server.go
@@ -12,12 +12,12 @@ import (
 	"strings"
 	"time"
 
+	metricsprovider "github.com/project-radius/radius/pkg/metrics/provider"
+	metricsservice "github.com/project-radius/radius/pkg/metrics/service"
+	metricsservicehostoptions "github.com/project-radius/radius/pkg/metrics/service/hostoptions"
 	"github.com/project-radius/radius/pkg/rp/kube"
 	"github.com/project-radius/radius/pkg/sdk"
-	metricsprovider "github.com/project-radius/radius/pkg/telemetry/metrics/provider"
-	metricsservice "github.com/project-radius/radius/pkg/telemetry/metrics/service"
-	metricsservicehostoptions "github.com/project-radius/radius/pkg/telemetry/metrics/service/hostoptions"
-	"github.com/project-radius/radius/pkg/telemetry/trace"
+	"github.com/project-radius/radius/pkg/trace"
 	"github.com/project-radius/radius/pkg/ucp/config"
 	"github.com/project-radius/radius/pkg/ucp/data"
 	"github.com/project-radius/radius/pkg/ucp/dataprovider"
